Give albums, the map and the two dead-end states a common scale
The album tile had a `Surface` behind its two lines of text, which made every tile a
card when its job is to show a picture and name it; the grid of twelve albums read as
twelve boxes. The label now sits on the screen's own background, and the "create album"
button — a stretched primary action above everything, spending 56 dp of a phone's
height on a control used once in a blue moon — moved into the heading of the section it
adds to. System tiles keep their icon because a category has no cover of its own; the
icon's disc went from 56 dp to 40 dp, which is what made those tiles look mostly empty.

Map pins were osmdroid's default teardrop, 44 dp of bitmap per photo, drawn beside a
text bubble for clusters — two scales of marker on one map. Both now come from
`setTextIcon`: a dot for a photo, the count for a cluster, one font size, the app's
accent. The strip behind them got a backing of its own and was lifted clear of the
attribution line, which is the one thing on that screen a provider's terms ask to keep
readable.

The two states that end in a dead end — Photos' permission gate, Cloud's failure — now
draw through PlaceholderScreen with an action, so the app says "allow access" and "try
again" in the same shape. Photos' own scan spinner matches.

Also the first CI pass's two errors: `Arrangement.spacedBy` wants Alignment.Vertical,
and `Alignment.Center` is neither Horizontal nor Vertical, so it is a compile error that
reads like a style choice.

+++ app/src/test/java/com/lumovault/app/domain/model/TimelineRailTest.kt
@@ -163,22 +163,36 @@ class TimelineRailTest {
         // Two months of equal weight split the rail in half, so their centres are a quarter of the way down
         // and three quarters. The pill is drawn at these numbers: a month named at fraction 0 would float
         // above the first tick it is supposed to be labelling, and one named at its last item would sit over
-        // the month below it.
+        // the month below it. (Doubles rather than floats because that is the `assertEquals` overload JUnit
+        // has had longest.)
         val months = TimelineRail.months(library)
 
-        assertEquals(0.25f, TimelineRail.fractionFor(months, months[0]), 0.001f)
-        assertEquals(0.75f, TimelineRail.fractionFor(months, months[1]), 0.001f)
+        assertEquals(0.25, TimelineRail.fractionFor(months, months[0]).toDouble(), 0.001)
+        assertEquals(0.75, TimelineRail.fractionFor(months, months[1]).toDouble(), 0.001)
     }
 
     @Test
     fun theCentreOfAWholeLibraryIsItsOwnMiddle() {
         val one = TimelineRail.months(listOf(day(epochDay = epochDay(2024, 3, 1), photos = 9)))
 
-        assertEquals("one month on the rail is the whole rail", 0.5f, TimelineRail.fractionFor(one, one[0]), 0.001f)
-        assertEquals("a month from a list it is not in names nothing",
-            0f, TimelineRail.fractionFor(one, RailMonth(YearMonth.of(2020, 1), firstItemIndex = 0, itemCount = 1)),
-            0.001f)
-        assertEquals(0f, TimelineRail.fractionFor(emptyList(), RailMonth(YearMonth.of(2024, 3), 0, 0)), 0.001f)
+        assertEquals(
+            "one month on the rail is the whole rail",
+            0.5,
+            TimelineRail.fractionFor(one, one[0]).toDouble(),
+            0.001,
+        )
+        assertEquals(
+            "a month from a list it is not in names nothing",
+            0.0,
+            TimelineRail.fractionFor(one, RailMonth(YearMonth.of(2020, 1), firstItemIndex = 0, itemCount = 1))
+                .toDouble(),
+            0.001,
+        )
+        assertEquals(
+            0.0,
+            TimelineRail.fractionFor(emptyList(), RailMonth(YearMonth.of(2024, 3), 0, 0)).toDouble(),
+            0.001,
+        )
     }
 
     @Test
